requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2024.8.30
click==8.1.7
dnspython==2.6.1
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.2
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==24.1
pillow==10.4.0
pydantic==2.9.2
pydantic-core==2.23.4
pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
replicate==1.0.2
rich==13.9.2
shellingham==1.5.4
sniffio==1.3.1
starlette==0.40.0
typer==0.12.5
typing-extensions==4.12.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
